# main: optimise "runtime" steps