--- a/_scripts/known_diffs/go1.11.13/badly_sorted_vendor_modules.txt.diff
+++ b/_scripts/known_diffs/go1.11.13/badly_sorted_vendor_modules.txt.diff
@@ -1,9 +1,9 @@
 diff --git b/vendor/modules.txt a/vendor/modules.txt
-index 49894ad..5b8d57c 100644
+index 01abe41..ebc5500 100644
 --- b/vendor/modules.txt
 +++ a/vendor/modules.txt
 @@ -1,16 +1,16 @@
- # github.com/rogpeppe/go-internal v1.3.0
+ # github.com/rogpeppe/go-internal v1.3.1
  github.com/rogpeppe/go-internal/cmd/txtar-addmod
 -github.com/rogpeppe/go-internal/module
 -github.com/rogpeppe/go-internal/txtar