--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,5 @@
 module github.com/myitcv/gobin
 
-require github.com/rogpeppe/go-internal v1.3.0
+require github.com/rogpeppe/go-internal v1.3.1
 
 go 1.11