… go.sum
… vendor/github.com/rogpeppe/go-internal/goproxytest/proxy.go
… vendor/modules.txt